+++ localbuild.ps1
@@ -1,4 +1,0 @@
-rm -force -recurse data/zerp-MelSkin
-deppth2 hpk -c BC7 -s .\textures\ -t .\data\zerp-MelSkin
-rm -force -recurse data/zerp-MelSkin 
-tcli build