+++ src/data.lua
@@ -65,14 +65,21 @@ mod.DressData = {
     ["Dark Side"] =
     {
         GrannyTexture = "Models/Melinoe/MelinoeTransform_Color",
+        ChildGrannyTexture = "Models/Melinoe/YoungMelTransform_Color",
     },
     ["Alternate Time"] =
     {
-        GrannyTexture = "zerp-MelSkin/skins/Alternate Time"
+        GrannyTexture = "zerp-MelSkin/Alternate Time",
+        ChildGrannyTexture = "zerp-MelSkin/YoungMelRed_Color",
+        TyphonRivalsPortraitMap = {
+            Portrait_Mel_Child_Defiant_01 = "Portrait_Melinoe_Child_Ending_Defiant_01",
+            Portrait_Mel_Child_Defiant_01_Exit = "Portrait_Melinoe_Child_Ending_Defiant_01_Exit",
+        },
+        DisableMelArmGlow = true,
     },
     Murderrrrr =
     {
-        GrannyTexture = "zerp-MelSkin/skins/Halloween 2025"
+        GrannyTexture = "zerp-MelSkin/Halloween 2025"
     },
     Void =
     {
@@ -83,6 +90,10 @@ mod.DressData = {
     None =
     {
         GrannyTexture = ""
+    },
+    Custom =
+    {
+        GrannyTexture = "zerp-MelSkinCustom/custom"
     }
 }
 
@@ -100,6 +111,19 @@ mod.DressDisplayOrder = {
     "Murderrrrr",
     "Void",
     "None",
+    "Custom"
+}
+
+mod.CustomDressDisplayOrder = {
+    "Lavender" ,
+    "Azure" ,
+    "Emerald" ,
+    "Onyx" ,
+    "Fuchsia" ,
+    "Gilded" ,
+    "Moonlight" ,
+    "Crimson" ,
+    "None"
 }
 
 -- for portraitprefix based on Arachne boon